refactor(schedule): tighten library surface
